add get_all_comments functionality
