+++ src/main.py
@@ -7,9 +7,9 @@ from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 from fastapi.middleware.cors import CORSMiddleware
 
-from apiclient.discovery import build
+from googleapiclient.discovery import build
 
-from v1.service import Comment, Transcript, Video, Ner
+from src.v1.service import Comment, Transcript, Video, Ner
 
 app = FastAPI()
 
@@ -95,11 +95,12 @@ async def sentiments_details(video_id: str):
 @app.get("/comments​/{video_id}")
 async def comments(video_id: str):
     comment = Comment(video_id)
+
     try:
-        # TODO: fix comment service
-        comments = jsonable_encoder(comment.get_list())
+        comments = jsonable_encoder(comment.get_all_comments())
     except:
         raise VideoException(video_id=video_id)
+
     return JSONResponse(content=comments)
 
 @app.get("/comments/{video_id}/controversial")
@@ -160,14 +161,13 @@ async def worldcloud(video_id: str):
 
 @app.get("/")
 async def root(request: Request):
-
     """
     Returns an object with live endpoints details.
-    
+
     To quickly look into the current live services from backend.
 
     Q. We already have /docs, /redoc, README.md and the swagger spec. So why is this required?
-    A. /docs, /redoc, README.md and swagger spec contains the 'mock' spec and won't tell the 
+    A. /docs, /redoc, README.md and swagger spec contains the 'mock' spec and won't tell the
        services which are live at the moment.
     """
 
@@ -229,4 +229,3 @@ async def root(request: Request):
             }
         }
     }
-
